lifting LogisticDrawer to campaignLogisticView parent
